[CodeGen] Introduce EraseDeadAllocAndStores pass.
It also refactors the implementation to Utils/.
