+++ compiler/src/iree/compiler/Codegen/Common/test/erase_dead_alloc_and_stores.mlir
@@ -1,4 +1,4 @@
-// RUN: iree-opt --iree-codegen-optimize-vector-transfer %s | FileCheck %s
+// RUN: iree-opt --iree-codegen-erase-dead-alloc-and-stores %s | FileCheck %s
 
 module {
   func.func @dead_alloc() {
